# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ApiEscolaDanca, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/usuarios": {
      "post": {
        "tags": [
          "ApiEscolaDanca, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Add_Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/usuarios/{userIdFirebase}": {
      "get": {
        "tags": [
          "ApiEscolaDanca, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userIdFirebase",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "ApiEscolaDanca, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userIdFirebase",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Update_Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "Add_Usuario": {
      "type": "object",
      "properties": {
        "userIdFirebase": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "cpf": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "dataNascimento": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Update_Usuario": {
      "type": "object",
      "properties": {
        "telefone": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}